Put all logic inside *component name* and use index just for exporting
